requirements.txt
Flask==3.*
SQLAlchemy
PyMySQL
gunicorn
cloud-sql-python-connector
functions-framework
requests
Werkzeug
graypy
python-dotenv
